Parser class added
- The Server::init() function creates a parser object and call Parser::parse(config_file)
- The file content is tokenized and the synthax is checked based on the tokens position
- when a Server or a Directive block is found, it is parsed and its infromations are stored into the Config
